sv/requirements.txt
requests==2.32.3
PyYAML==6.0.2
garminconnect>=0.2.25
